refactor(package): Refactored the file structure of the repository, added CI, and renamed the package to spac.
