+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!--next-version-placeholder-->
 
+## v0.1.0 (5/4/2023)
+
+- Renamed the workflow to `spac` and reorganized the directory.
+
 ## v0.1.0 (29/11/2022)
 
 - First release of `scworkflow`!
