+++ pyproject.toml
@@ -1,8 +1,8 @@
 [tool.poetry]
-name = "scworkflow"
+name = "spac"
 version = "0.1.0"
 description = "A package for various single cell MxIF/CODEX analysis steps. "
-authors = ["George Zaki"]
+authors = ["George Zaki", "Rui He", "Fang Liu"]
 license = "BSD 3-Clause"
 readme = "README.md"
