+++ tests/test_rename_clustering.py
@@ -2,7 +2,7 @@ import sys
 import os
 sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/../src")
 
-from scsaworkflow.transformations import rename_clustering
+from spac.transformations import rename_clustering
 import anndata
 import pandas as pd
 import numpy as np
